requirements.txt
asgiref==3.6.0
astroid==2.15.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.6
dj-database-url==2.0.0
dj-rest-auth==3.0.0
Django==3.2
django-allauth==0.54.0
django-colorfield==0.8.0
django-dotenv==1.4.2
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-import-export==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
filelock==3.12.2
fsspec==2023.6.0
gunicorn==20.1.0
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycparser==2.21
pydot==1.4.2
PyJWT==1.7.1
pylint==2.17.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tablib==3.5.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
